[open][upstream][build][baseline] Snapshot of upstream chromium v72 stable (tag: 72.0.3626.121)
:Release Notes:
Base upstream v72 stable

:Detailed Notes:
Snapshot of upstream/chromium:
    Repo: https://chromium.googlesource.com/chromium/src.git
    commit ec19c835273c380b90a9f1fe2ffb2589c0664da6 (tag: 72.0.3626.121, origin/72.0.3626.121)
    Author: chrome-release-bot <chrome-release-bot@chromium.org>
    Date:   Thu Feb 28 12:44:19 2019 -0800

        Publish DEPS for 72.0.3626.121

Snapshot of upstream/depot_tools:
    Repo: https://chromium.googlesource.com/chromium/tools/depot_tools.git
    commit f5f2ef9c938e4c46b25b43c754563a7a39700c15 (HEAD -> master, origin/master, origin/HEAD)
    Author: recipe-roller <recipe-mega-autoroller@chops-service-accounts.iam.gserviceaccount.com>
    Date:   Mon Jun 3 18:16:55 2019 +0000

        Roll recipe dependencies (trivial)

Plus:
    Remove src/v8 - split into another repo
    Run scripts/build/remove-nonessential-files.sh

Steps followed:
1. Clone neva chromium
    git clone neva/chromium chromium-neva-prep/src
2. Clone depot_tools
    git clone
    https://chromium.googlesource.com/chromium/tools/depot_tools.git
    chromium-neva-prep/depot_tools
3. Checkout chromium upstream baseline (eg: v72-stable)
    cd chromium-neva-prep/src
    git checkout -b upstream-chr72-stable ec19c835273c
4. Run gclient to download third_party sources
    export GYP_CHROMIUM_NO_ACTION=1
    export PATH="<path-to-depot_tools>;$PATH"
    gclient config  --name=src --unmanaged https://chromium.googlesource.com/chromium/src.git
    gclient sync --with_tags --with_branch_heads
5. Run script to remove non-essential files
    ./scripts/build/remove-nonessential-files.sh
6. Run script to add essential files
    ./scripts/build/copyback-essential-files.sh
7. Init git repo
    git init
8. Add all files to git commit
    git add -Af *
    git add .gclient*
    git commit -a

Initial size:
chromium-neva-prep$ du -sh .
6.3G	.

:Testing Performed:
Check size is acceptable

:QA Notes:
Test-cases: N/A

:Issues Addressed:
[PLAT-79872] Create RP fork of Neva Chromium

+++ src/build/android/gyp/util/resource_utils.py
@@ -25,6 +25,23 @@ EMPTY_ANDROID_MANIFEST_PATH = os.path.join(
     _SOURCE_ROOT, 'build', 'android', 'AndroidManifest.xml')
 
 
+# A variation of this lists also exists in:
+# //base/android/java/src/org/chromium/base/LocaleUtils.java
+# //ui/android/java/src/org/chromium/base/LocalizationUtils.java
+CHROME_TO_ANDROID_LOCALE_MAP = {
+    'en-GB': 'en-rGB',
+    'en-US': 'en-rUS',
+    'es-419': 'es-rUS',
+    'fil': 'tl',
+    'he': 'iw',
+    'id': 'in',
+    'pt-PT': 'pt-rPT',
+    'pt-BR': 'pt-rBR',
+    'yi': 'ji',
+    'zh-CN': 'zh-rCN',
+    'zh-TW': 'zh-rTW',
+}
+
 # Represents a line from a R.txt file.
 _TextSymbolEntry = collections.namedtuple('RTextEntry',
     ('java_type', 'resource_type', 'name', 'value'))
@@ -197,7 +214,7 @@ def CreateRJavaFiles(srcjar_dir, package, main_r_txt_file,
   packages = list(extra_res_packages)
   r_txt_files = list(extra_r_txt_files)
 
-  if package not in packages:
+  if package and package not in packages:
     # Sometimes, an apk target and a resources target share the same
     # AndroidManifest.xml and thus |package| will already be in |packages|.
     packages.append(package)
@@ -312,7 +329,11 @@ public final class R {
         public static final {{ e.java_type }} {{ e.name }} = {{ e.value }};
         {% endfor %}
         {% for e in non_final_resources[resource_type] %}
+            {% if e.value != '0' %}
         public static {{ e.java_type }} {{ e.name }} = {{ e.value }};
+            {% else %}
+        public static {{ e.java_type }} {{ e.name }};
+            {% endif %}
         {% endfor %}
     }
     {% endfor %}
@@ -386,11 +407,22 @@ def ExtractDeps(dep_zips, deps_dir):
 
 
 class _ResourceBuildContext(object):
-  """A temporary directory for packaging and compiling Android resources."""
-  def __init__(self):
+  """A temporary directory for packaging and compiling Android resources.
+
+  Args:
+    temp_dir: Optional root build directory path. If None, a temporary
+      directory will be created, and removed in Close().
+  """
+  def __init__(self, temp_dir=None):
     """Initialized the context."""
     # The top-level temporary directory.
-    self.temp_dir = tempfile.mkdtemp()
+    if temp_dir:
+      self.temp_dir = temp_dir
+      self.remove_on_exit = False
+    else:
+      self.temp_dir = tempfile.mkdtemp()
+      self.remove_on_exit = True
+
     # A location to store resources extracted form dependency zip files.
     self.deps_dir = os.path.join(self.temp_dir, 'deps')
     os.mkdir(self.deps_dir)
@@ -405,14 +437,15 @@ class _ResourceBuildContext(object):
 
   def Close(self):
     """Close the context and destroy all temporary files."""
-    shutil.rmtree(self.temp_dir)
+    if self.remove_on_exit:
+      shutil.rmtree(self.temp_dir)
 
 
 @contextlib.contextmanager
-def BuildContext():
+def BuildContext(temp_dir=None):
   """Generator for a _ResourceBuildContext instance."""
   try:
-    context = _ResourceBuildContext()
+    context = _ResourceBuildContext(temp_dir)
     yield context
   finally:
     context.Close()
@@ -432,11 +465,16 @@ def ResourceArgsParser():
 
   build_utils.AddDepfileOption(output_opts)
 
-  input_opts.add_argument('--android-sdk-jars', required=True,
-                        help='Path to the android.jar file.')
+  input_opts.add_argument('--include-resources', required=True, action="append",
+                        help='Paths to arsc resource files used to link '
+                             'against. Can be specified multiple times.')
 
   input_opts.add_argument('--aapt-path', required=True,
                          help='Path to the Android aapt tool')
+
+  input_opts.add_argument('--aapt2-path',
+                          help='Path to the Android aapt2 tool. If in different'
+                          ' directory from --aapt-path.')
 
   input_opts.add_argument('--dependencies-res-zips', required=True,
                     help='Resources zip archives from dependents. Required to '
@@ -468,7 +506,11 @@ def HandleCommonOptions(options):
     options: the result of parse_args() on the parser returned by
         ResourceArgsParser(). This function updates a few common fields.
   """
-  options.android_sdk_jars = build_utils.ParseGnList(options.android_sdk_jars)
+  options.include_resources = [build_utils.ParseGnList(r) for r in
+                               options.include_resources]
+  # Flatten list of include resources list to make it easier to use.
+  options.include_resources = [r for resources in options.include_resources
+                               for r in resources]
 
   options.dependencies_res_zips = (
       build_utils.ParseGnList(options.dependencies_res_zips))
@@ -485,3 +527,6 @@ def HandleCommonOptions(options):
         build_utils.ParseGnList(options.extra_r_text_files))
   else:
     options.extra_r_text_files = []
+
+  if not options.aapt2_path:
+    options.aapt2_path = options.aapt_path + '2'
