+++ src/components/sync/device_info/device_info_sync_bridge.cc
@@ -87,7 +87,8 @@ DeviceInfoSyncBridge::DeviceInfoSyncBridge(
     OnceModelTypeStoreFactory store_factory,
     std::unique_ptr<ModelTypeChangeProcessor> change_processor)
     : ModelTypeSyncBridge(std::move(change_processor)),
-      local_device_info_provider_(local_device_info_provider) {
+      local_device_info_provider_(local_device_info_provider),
+      weak_ptr_factory_(this) {
   DCHECK(local_device_info_provider);
 
   // This is not threadsafe, but presuably the provider initializes on the same
@@ -102,7 +103,7 @@ DeviceInfoSyncBridge::DeviceInfoSyncBridge(
 
   std::move(store_factory)
       .Run(DEVICE_INFO, base::BindOnce(&DeviceInfoSyncBridge::OnStoreCreated,
-                                       base::AsWeakPtr(this)));
+                                       weak_ptr_factory_.GetWeakPtr()));
 }
 
 DeviceInfoSyncBridge::~DeviceInfoSyncBridge() {}
@@ -119,11 +120,9 @@ base::Optional<ModelError> DeviceInfoSyncBridge::MergeSyncData(
   DCHECK(change_processor()->IsTrackingMetadata());
   const DeviceInfo* local_info =
       local_device_info_provider_->GetLocalDeviceInfo();
-  // If our dependency was yanked out from beneath us, we cannot correctly
-  // handle this request, and all our data will be deleted soon.
-  if (local_info == nullptr) {
-    return {};
-  }
+  // DEVICE_INFO sync is running; DeviceInfo thus exists (it gets cleared only
+  // synchronously with disabling DEVICE_INFO sync).
+  DCHECK(local_info);
 
   // Local data should typically be near empty, with the only possible value
   // corresponding to this device. This is because on signout all device info
@@ -172,11 +171,9 @@ base::Optional<ModelError> DeviceInfoSyncBridge::ApplySyncChanges(
   DCHECK(has_provider_initialized_);
   const DeviceInfo* local_info =
       local_device_info_provider_->GetLocalDeviceInfo();
-  // If our dependency was yanked out from beneath us, we cannot correctly
-  // handle this request, and all our data will be deleted soon.
-  if (local_info == nullptr) {
-    return {};
-  }
+  // DEVICE_INFO sync is running; DeviceInfo thus exists (it gets cleared only
+  // synchronously with disabling DEVICE_INFO sync).
+  DCHECK(local_info);
 
   std::unique_ptr<WriteBatch> batch = store_->CreateWriteBatch();
   bool has_changes = false;
@@ -218,7 +215,7 @@ void DeviceInfoSyncBridge::GetData(StorageKeyList storage_keys,
   std::move(callback).Run(std::move(batch));
 }
 
-void DeviceInfoSyncBridge::GetAllData(DataCallback callback) {
+void DeviceInfoSyncBridge::GetAllDataForDebugging(DataCallback callback) {
   auto batch = std::make_unique<MutableDataBatch>();
   for (const auto& kv : all_data_) {
     batch->Put(kv.first, CopyToEntityData(*kv.second));
@@ -236,20 +233,22 @@ std::string DeviceInfoSyncBridge::GetStorageKey(const EntityData& entity_data) {
   return entity_data.specifics.device_info().cache_guid();
 }
 
-ModelTypeSyncBridge::DisableSyncResponse
-DeviceInfoSyncBridge::ApplyDisableSyncChanges(
+ModelTypeSyncBridge::StopSyncResponse
+DeviceInfoSyncBridge::ApplyStopSyncChanges(
     std::unique_ptr<MetadataChangeList> delete_metadata_change_list) {
   // TODO(skym, crbug.com/659263): Would it be reasonable to pulse_timer_.Stop()
   // or subscription_.reset() here?
 
   // Remove all local data, if sync is being disabled, the user has expressed
   // their desire to not have knowledge about other devices.
-  store_->DeleteAllDataAndMetadata(base::DoNothing());
-  if (!all_data_.empty()) {
-    all_data_.clear();
-    NotifyObservers();
+  if (delete_metadata_change_list) {
+    store_->DeleteAllDataAndMetadata(base::DoNothing());
+    if (!all_data_.empty()) {
+      all_data_.clear();
+      NotifyObservers();
+    }
   }
-  return DisableSyncResponse::kModelStillReadyToSync;
+  return StopSyncResponse::kModelStillReadyToSync;
 }
 
 bool DeviceInfoSyncBridge::IsSyncing() const {
@@ -268,8 +267,7 @@ std::unique_ptr<DeviceInfo> DeviceInfoSyncBridge::GetDeviceInfo(
 std::vector<std::unique_ptr<DeviceInfo>>
 DeviceInfoSyncBridge::GetAllDeviceInfo() const {
   std::vector<std::unique_ptr<DeviceInfo>> list;
-  for (ClientIdToSpecifics::const_iterator iter = all_data_.begin();
-       iter != all_data_.end(); ++iter) {
+  for (auto iter = all_data_.begin(); iter != all_data_.end(); ++iter) {
     list.push_back(SpecificsToModel(*iter->second));
   }
   return list;
@@ -350,7 +348,7 @@ void DeviceInfoSyncBridge::OnStoreCreated(
 
   store_ = std::move(store);
   store_->ReadAllData(base::BindOnce(&DeviceInfoSyncBridge::OnReadAllData,
-                                     base::AsWeakPtr(this)));
+                                     weak_ptr_factory_.GetWeakPtr()));
 }
 
 void DeviceInfoSyncBridge::OnReadAllData(
@@ -379,8 +377,9 @@ void DeviceInfoSyncBridge::OnReadAllData(
 
 void DeviceInfoSyncBridge::LoadMetadataIfReady() {
   if (has_data_loaded_ && has_provider_initialized_) {
-    store_->ReadAllMetadata(base::BindOnce(
-        &DeviceInfoSyncBridge::OnReadAllMetadata, base::AsWeakPtr(this)));
+    store_->ReadAllMetadata(
+        base::BindOnce(&DeviceInfoSyncBridge::OnReadAllMetadata,
+                       weak_ptr_factory_.GetWeakPtr()));
   }
 }
 
@@ -470,9 +469,9 @@ void DeviceInfoSyncBridge::SendLocalData() {
 
 void DeviceInfoSyncBridge::CommitAndNotify(std::unique_ptr<WriteBatch> batch,
                                            bool should_notify) {
-  store_->CommitWriteBatch(
-      std::move(batch),
-      base::BindOnce(&DeviceInfoSyncBridge::OnCommit, base::AsWeakPtr(this)));
+  store_->CommitWriteBatch(std::move(batch),
+                           base::BindOnce(&DeviceInfoSyncBridge::OnCommit,
+                                          weak_ptr_factory_.GetWeakPtr()));
   if (should_notify) {
     NotifyObservers();
   }
